refactor: extract executeMotions function
