+++ src/input.ts
@@ -1,19 +1,13 @@
 import { isLeft } from "fp-ts/lib/Either";
-import { isSome, none, some } from "fp-ts/lib/Option";
-import {
-  commands,
-  Disposable,
-  InputBox,
-  Selection,
-  TextEditor,
-  window,
-} from "vscode";
+import { isSome, none } from "fp-ts/lib/Option";
+import { commands, Disposable, Selection, TextEditor, window } from "vscode";
 import {
   HierarchicalDisposer,
   withChildDisposer,
 } from "./hierarchical-disposer";
 import { Highlighter } from "./highlight";
 import { parseVimMotions } from "./motions";
+import { executeMotions } from "./motions/execute";
 
 export async function processVimMotionInput({
   disposer: parentDisposer,
@@ -41,7 +35,6 @@ export async function processVimMotionInput({
 class VimMotionInput implements Disposable {
   private readonly inputBox = window.createInputBox();
   private readonly highlighter: Highlighter;
-  private readonly restoreSelection: () => void;
   private readonly donePromise: Promise<boolean>;
 
   public dispose: Disposable["dispose"];
@@ -65,19 +58,11 @@ class VimMotionInput implements Disposable {
 
     this.highlighter = highlighter;
 
-    this.restoreSelection = () => {
-      editor.selection = initialSelection;
-      this.highlighter.highlight();
-    };
-
     this.donePromise = new Promise<boolean>((resolve) => {
       let accepted = false;
 
       disposer.add(
         this.inputBox.onDidHide(() => {
-          if (!accepted) {
-            this.restoreSelection();
-          }
           resolve(accepted);
         }),
       );
@@ -104,6 +89,7 @@ class VimMotionInput implements Disposable {
   private parseAndExecuteMotions = async (input: string) => {
     const result = parseVimMotions(input);
     this.inputBox.validationMessage = undefined;
+    this.highlighter.highlight();
 
     if (isLeft(result)) {
       if (isSome(result.left)) {
@@ -113,16 +99,7 @@ class VimMotionInput implements Disposable {
       return;
     }
 
-    this.restoreSelection();
-    let lastCommand: Thenable<unknown> = Promise.resolve();
-
-    for (const motion of result.right.motion) {
-      lastCommand = commands.executeCommand("cursorMove", {
-        to: motion.direction,
-        value: motion.lines,
-      });
-    }
-
-    await lastCommand;
+    await executeMotions(result.right.motion);
+    this.inputBox.value = result.right.unmatchedInput;
   };
 }
